add filter service

+++ Services/WorkerService/WorkerService.DAL/DTO/ActivityDtos.cs
@@ -4,8 +4,8 @@ namespace WorkerService.DAL.DTO;
 
 public class ActivityDtos
 {
-    [Required] public ulong id { get; set; }
+    [Required] public int id { get; set; }
     [Required] public string direction { get; set; } = null!;
     [Required] public string type { get; set; } = null!;
-    [Required] public ulong filter_id { get; set; }
+    [Required] public int filter_id { get; set; }
 }
