+++ Services/WorkerService/WorkerService.DAL/Сontracts/CreateResume.cs
@@ -3,8 +3,8 @@ namespace WorkerService.DAL.Contracts;
 public record CreateResume(
     string skill,
     string? city,
-    short? experience,
-    ulong? education_id,
+    int? experience,
+    int? education_id,
     int? wantedSalary,
     string post
 );
